+++ src/eip7594/eip7594.c
@@ -367,12 +367,12 @@ static void deduplicate_commitments(
 C_KZG_RET compute_verify_cell_kzg_proof_batch_challenge(
     fr_t *challenge_out,
     const Bytes48 *commitments_bytes,
-    size_t num_commitments,
+    uint64_t num_commitments,
     const uint64_t *commitment_indices,
     const uint64_t *cell_indices,
     const Cell *cells,
     const Bytes48 *proofs_bytes,
-    size_t num_cells
+    uint64_t num_cells
 ) {
     C_KZG_RET ret;
     uint8_t *bytes = NULL;
